+++ iron-spi/src/main/java/io/axway/iron/spi/storage/SnapshotStore.java
@@ -1,6 +1,7 @@
 package io.axway.iron.spi.storage;
 
 import java.io.*;
+import java.math.BigInteger;
 import java.util.*;
 
 /**
@@ -14,7 +15,7 @@ public interface SnapshotStore {
      * @return the {@code OutputStream} to be used to write the snapshot content.
      * @throws IOException in case of error when trying to provide access to the {@code OutputStream}
      */
-    OutputStream createSnapshotWriter(long transactionId) throws IOException;
+    OutputStream createSnapshotWriter(BigInteger transactionId) throws IOException;
 
     /**
      * Retrieve an existing snapshot in the store.
@@ -23,19 +24,19 @@ public interface SnapshotStore {
      * @return the {@code InputStream} to be used to read the snapshot content.
      * @throws IOException in case of error when trying to provide access to the {@code InputStream}
      */
-    InputStream createSnapshotReader(long transactionId) throws IOException;
+    InputStream createSnapshotReader(BigInteger transactionId) throws IOException;
 
     /**
      * List all the existing snapshot.
      *
      * @return the list of snapshot transactions id.
      */
-    List<Long> listSnapshots();
+    List<BigInteger> listSnapshots();
 
     /**
      * Delete a snapshot.
      *
      * @param transactionId the transaction id of the snapshot to be deleted.
      */
-    void deleteSnapshot(long transactionId);
+    void deleteSnapshot(BigInteger transactionId);
 }
